Stub out Note API
Introduce wreq for Http Client library.

Probably `putStrLn $ show` is not the right thing to do.

Need to handle HTTP exceptions.

Need to test with the service actually running. (Only compiled and
tested with the service down.)

Need to remove the scaffold code.

Need to figure out how/what to unit test.